fix: avoid to launch task manager in taskrunner #1080
