+++ datashare-app/src/main/java/org/icij/datashare/WebApp.java
@@ -1,12 +1,16 @@
 package org.icij.datashare;
 
 import net.codestory.http.WebServer;
+import org.icij.datashare.batch.BatchSearch;
+import org.icij.datashare.batch.BatchSearchRepository;
 import org.icij.datashare.cli.DatashareCli;
 import org.icij.datashare.cli.Mode;
 import org.icij.datashare.cli.QueueType;
 import org.icij.datashare.com.bus.amqp.QpidAmqpServer;
 import org.icij.datashare.mode.CommonMode;
+import org.icij.datashare.tasks.BatchSearchRunner;
 import org.icij.datashare.tasks.TaskFactory;
+import org.icij.datashare.tasks.TaskManager;
 
 import java.awt.*;
 import java.io.IOException;
@@ -48,6 +52,7 @@ public class WebApp {
             waitForServerToBeUp(parseInt(mode.properties().getProperty(PropertiesProvider.TCP_LISTEN_PORT)));
             Desktop.getDesktop().browse(URI.create(new URI("http://localhost:")+mode.properties().getProperty(PropertiesProvider.TCP_LISTEN_PORT)));
         }
+        requeueDatabaseBatches(mode.get(BatchSearchRepository.class), mode.get(TaskManager.class));
         webServerThread.join();
     }
 
@@ -62,6 +67,13 @@ public class WebApp {
            } else {
                Thread.sleep(500);
            }
+        }
+    }
+
+    private static void requeueDatabaseBatches(BatchSearchRepository repository, TaskManager taskManager) throws IOException {
+        for (String batchSearchUuid: repository.getQueued()) {
+            BatchSearch batchSearch = repository.get(batchSearchUuid);
+            taskManager.startTask(batchSearchUuid, BatchSearchRunner.class.getName(), batchSearch.user);
         }
     }
