+++ datashare-cli/src/main/java/org/icij/datashare/cli/Mode.java
@@ -1,7 +1,7 @@
 package org.icij.datashare.cli;
 
 public enum Mode {
-    LOCAL(true), SERVER(true), CLI(false), NER(true), BATCH_SEARCH(false), BATCH_DOWNLOAD(false), EMBEDDED(true);
+    LOCAL(true), SERVER(true), CLI(false), NER(true), TASK_RUNNER(false), EMBEDDED(true);
     private final boolean webServer;
 
     Mode(boolean webServer) {
